+++ crates/plugin_script/tests/mod.rs
@@ -118,11 +118,8 @@ fn load_parse_and_analyze_deps() {
         ]
       );
 
-      let mut resource_pot = ResourcePot::new(
-        ResourcePotId::new("index".to_string()),
-        ResourcePotType::Js,
-        "any".into(),
-      );
+      let mut resource_pot =
+        ResourcePot::new(ResourcePotId::new("index".to_string()), ResourcePotType::Js);
 
       resource_pot.resource_pot_type = ResourcePotType::Js;
       resource_pot.meta = ResourcePotMetaData::Js(JsResourcePotMetaData {